Refactored DataSctructure and StacksAndQueues file to be within challenge folder and deleted unnecessary folders. Correcting code challenges based on Ashlyns comments on remaining code challenges.
