--- challenges/DataStructures/src/main/java/DataStructures/linkedList/Node.java
@@ -1,12 +1,0 @@
-package DataStructures.linkedList;
-
-public class Node {
-    public int value;
-    public Node head;
-    public Node next;
-
-    Node(int value) {
-        this.value = value;
-    }
-
-}
