--- challenges/DataStructures/src/test/java/DataStructures/LibraryTest.java
@@ -1,15 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package DataStructures;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class LibraryTest {
-    @Test public void testSomeLibraryMethod() {
-        Library classUnderTest = new Library();
-        assertTrue("someLibraryMethod should return 'true'", classUnderTest.someLibraryMethod());
-    }
-}
-
