--- challenges/StacksAndQueues/src/main/java/stacksandqueues/Library.java
@@ -1,8 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package stacksandqueues;
-
-public class Library {
-
-}
