--- challenges/StacksAndQueues/src/main/java/stacksandqueues/Node.java
@@ -1,13 +1,0 @@
-package stacksandqueues;
-
-public class Node {
-    public int value;
-    public Node next;
-    public Node front;
-    public Node rear;
-
-    public Node(int value) {
-        this.value = value;
-    }
-
-}
