--- challenges/StacksAndQueues/src/test/java/stacksandqueues/LibraryTest.java
@@ -1,11 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package stacksandqueues;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class LibraryTest {
-
-}
