+++ challenges/src/main/java/challenges/DataStructures/Library.java
@@ -1,0 +1,39 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package challenges.DataStructures;
+
+public class Library {
+    public boolean someLibraryMethod() {
+        return true;
+    }
+}
+
+//    Node head = null;
+//    Node tail = null;
+//
+//    public void insert(String value) {
+//        Node node = new Node(value);
+//        if(head == null) {
+//            head = node;
+//            tail = node;
+//        } else {
+//            node.next = head;
+//            head = node;
+//        }
+//    }
+//
+//    public String toString() {
+//        String output = "";
+//        Node currentNodeIAmLookingAt = head;
+//        while(currentNodeIAmLookingAt != null) {
+//            output = output + System.out.printf("{%s} -> ", currentNodeIAmLookingAt.value);
+//            currentNodeIAmLookingAt = currentNodeIAmLookingAt.next;
+//        }
+//        output = output + "null";
+//        return output;
+//    }
+
+//    public static boolean includes(String value) {
+//
+//    }
