+++ challenges/src/main/java/challenges/linkedList/Node.java
@@ -1,0 +1,12 @@
+package challenges.linkedList;
+
+public class Node {
+    public int value;
+    public Node head;
+    public Node next;
+
+    Node(int value) {
+        this.value = value;
+    }
+
+}
