+++ challenges/src/main/java/challenges/stacksandqueues/Library.java
@@ -1,0 +1,8 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package challenges.stacksandqueues;
+
+public class Library {
+
+}
