+++ challenges/src/main/java/challenges/stacksandqueues/Node.java
@@ -1,0 +1,13 @@
+package challenges.stacksandqueues;
+
+public class Node {
+    public int value;
+    public Node next;
+    public Node front;
+    public Node rear;
+
+    public Node(int value) {
+        this.value = value;
+    }
+
+}
