+++ challenges/src/test/java/challenges/stacksandqueues/LibraryTest.java
@@ -1,0 +1,11 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package challenges.stacksandqueues;
+
+import org.junit.Test;
+import static org.junit.Assert.*;
+
+public class LibraryTest {
+
+}
